Fix bug in ChallengeComponent
Bug was caused by attempting to access methods of null Challenge object
in cases where the challenge was not found. Fixed it by adding checks
and accessing the methods only if the challenge was found.

Had to change some of the tests for the component as well, and refactored
them a bit while I was at it.