Update social auth service to work with JWT
Had to change the way the 'authenticated' view handles passing the data to
the main window. Using a callback wasn't working, so changed it to use
the window.postMessage method to send the data. The social auth service
then listens for messages from the popup and sets the token and user
profile from the sent data.